renamed files and folders for consistency
