+++ src/reentry/trajectories/ballistic.py
