+++ src/reentry/trajectories/launch_vehicle.py
