+++ src/tradeoff/main.py
@@ -1,4 +1,4 @@
-from tradeOff import TradeOff
+from tradeoff import TradeOff
 import time
